Bug fixing and small improvments.
Now instead a single result at the getting user from database uses get
result list. It is mades, because getSingleResult may throw exception
NoSuchException. Now it works next way - get list of result, check that
is not empty and take first element from the list.

Also removed Component annotaions because it is not clear to
understanding. Spring create this bean, but Repository annotations give
more clear undrestanding of the what is the class.